0.10.2

+++ packages/lucia-sveltekit/src/auth/session.ts
@@ -81,14 +81,14 @@ export const deleteExpiredUserSessionsFunction = (context: Context) => {
         const currentTime = new Date().getTime();
         const expiredUserAccessTokens = sessions
             .filter((val) => val.expires < currentTime)
-            .map((val) => val.accessToken);
+            .map((val) => val.access_token);
+        if (expiredUserAccessTokens.length === 0) return;
         await context.adapter.deleteSessionByAccessToken(
             ...expiredUserAccessTokens
         );
     };
     return deleteExpiredUserSessions;
 };
-
 
 type RefreshSession = (refreshToken: string) => Promise<{
     session: Session;
